+++ Source/MediaStorageAndFileFormat/gdcmJPEGLSCodec.cxx
@@ -411,7 +411,7 @@ void JPEGLSCodec::SetLossyError(int error)
   LossyError = error;
 }
 
-bool JPEGLSCodec::Decode(DataElement const &, char* , uint32_t ,
+bool JPEGLSCodec::Decode(DataElement const &, char* , size_t,
               uint32_t , uint32_t , uint32_t ,
               uint32_t , uint32_t , uint32_t )
 {
